Change ponly fitter changes to reflect believe of the analysis
PP radius is now used for the calculation of the feeddown CFs. Proton Lambda is calculated
with Lednicky NLO (systematic variation to LO) and Proton Xi is calculated by
HAL QCD Potential. The baseline in the default now only contains a constant, and as a systematic
a slope is added. The fitting range is now limited to an area where we believe pp to be flat.